chore(preview): align code style and naming, improve partitioning, fix MobX warning
