+++ src/electron/main.ts
@@ -24,13 +24,13 @@ Fs.writeFileSync(Path.join(__dirname, 'app.html'), RENDERER_DOCUMENT);
 
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
-let win: BrowserWindow | undefined;
+let window: BrowserWindow | undefined;
 
 async function createWindow(): Promise<void> {
 	const { width = 1280, height = 800 } = screen.getPrimaryDisplay().workAreaSize;
 
 	// Create the browser window.
-	win = new BrowserWindow({
+	window = new BrowserWindow({
 		width,
 		height,
 		minWidth: 780,
@@ -41,7 +41,7 @@ async function createWindow(): Promise<void> {
 	});
 
 	// and load the index.html of the app.
-	win.loadURL(
+	window.loadURL(
 		Url.format({
 			pathname: Path.join(__dirname, 'app.html'),
 			protocol: 'file:',
@@ -56,8 +56,8 @@ async function createWindow(): Promise<void> {
 
 	// tslint:disable-next-line:no-any
 	const send = (message: ServerMessage) => {
-		if (win) {
-			win.webContents.send('message', message);
+		if (window) {
+			window.webContents.send('message', message);
 		}
 	};
 
@@ -80,7 +80,7 @@ async function createWindow(): Promise<void> {
 		}
 	});
 
-	server.on('client-message', (envelope: string) => {
+	server.on('preview-message', (envelope: string) => {
 		try {
 			const message = JSON.parse(envelope);
 
@@ -130,15 +130,15 @@ async function createWindow(): Promise<void> {
 	// win.webContents.openDevTools();
 
 	// Emitted when the window is closed.
-	win.on('closed', () => {
+	window.on('closed', () => {
 		// Dereference the window object, usually you would store windows
 		// in an array if your app supports multi windows, this is the time
 		// when you should delete the corresponding element.
-		win = undefined;
+		window = undefined;
 	});
 
 	// Disable navigation on the host window object, triggered by system drag and drop
-	win.webContents.on('will-navigate', e => {
+	window.webContents.on('will-navigate', e => {
 		e.preventDefault();
 	});
 
@@ -157,7 +157,7 @@ async function createWindow(): Promise<void> {
 				console.warn('An error occurred: ', err);
 			});
 	}
-	checkForUpdates(win);
+	checkForUpdates(window);
 }
 
 const log = require('electron-log');
@@ -182,13 +182,13 @@ app.on('window-all-closed', () => {
 app.on('activate', async () => {
 	// On macOS it's common to re-create a window in the app when the
 	// dock icon is clicked and there are no other windows open.
-	if (!win) {
+	if (!window) {
 		await createWindow();
 	}
 });
 
 ipcMain.on('request-check-for-updates', () => {
-	if (win) {
-		checkForUpdates(win, true);
+	if (window) {
+		checkForUpdates(window, true);
 	}
 });
