+++ src/preview/document.ts
@@ -82,7 +82,6 @@ export const previewDocument = `<!doctype html>
 		</div>
 	</div>
 	<script src="/scripts/vendor.js" data-script="vendor"></script>
-	<script src="/scripts/renderer.js" data-script="renderer"></script>
 	<script src="/scripts/components.js" data-script="components"></script>
 	<script src="/scripts/preview.js" data-script="preview"></script>
 </body>
