+++ src/preview/error-boundary.tsx
@@ -1,0 +1,29 @@
+import { ErrorMessage } from './error-message';
+import * as React from 'react';
+
+export interface ErrorBoundaryProps {
+	name: string;
+}
+
+export interface ErrorBoundaryState {
+	errorMessage?: string;
+}
+
+export class ErrorBoundary extends React.Component<ErrorBoundaryProps, ErrorBoundaryState> {
+	public state = {
+		errorMessage: ''
+	};
+
+	public componentDidCatch(error: Error): void {
+		this.setState({
+			errorMessage: error.message
+		});
+	}
+
+	public render(): JSX.Element {
+		if (this.state.errorMessage) {
+			return <ErrorMessage patternName={this.props.name} error={this.state.errorMessage} />;
+		}
+		return this.props.children as JSX.Element;
+	}
+}
