+++ src/preview/error-message.tsx
@@ -1,0 +1,29 @@
+import * as React from 'react';
+
+export interface ErrorMessageProps {
+	error: string;
+	patternName: string;
+}
+
+export const ErrorMessage: React.StatelessComponent<ErrorMessageProps> = props => (
+	<div
+		style={{
+			backgroundColor: 'rgb(240, 40, 110)',
+			color: 'white',
+			padding: '12px 15px',
+			textAlign: 'center'
+		}}
+	>
+		<p
+			style={{
+				margin: '0',
+				fontFamily:
+					'-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+				fontSize: '15px',
+				lineHeight: '22px'
+			}}
+		>
+			{`<${props.patternName}/> failed to render: ${props.error}`}
+		</p>
+	</div>
+);
