+++ src/preview/store/highlight-area.ts
