+++ src/preview/store/page.ts
@@ -1,0 +1,6 @@
+import { PageElement } from './page-element';
+
+export interface Page {
+	id: string;
+	root: PageElement;
+}
